fixed javadoc issue
