+++ java/src/test/java/com/amido/stacks/azure/servicebus/publisher/TestApplicationEvent.java
@@ -1,4 +1,4 @@
-package com.amido.stacks.core.azure.servicebus.publisher;
+package com.amido.stacks.azure.servicebus.publisher;
 
 import com.amido.stacks.core.messaging.event.ApplicationEvent;
